chore(cli): Phase 1 cleanup — consolidate tests, drop dead barrel
- Move all *.test.ts and e2e-testkit.ts into packages/cli/src/test/
  (mirroring source structure under src/test/bin/ where relevant) so
  production code and tests stop sharing the top-level directory.
- Delete the unused src/bin/lib.ts barrel; inline its imports.
- Remove leftover no-op { ... } blocks in cmdMigrate and cmdCheck.
- Hoist the core-command list in chkit.ts so it is defined once.
- Update tsconfig exclude to src/test/**.
