+++ packages/cli/src/bin/commands/migrate.ts
@@ -111,229 +111,227 @@ async function cmdMigrate(runCtx: CommandRunContext): Promise<void> {
   }
   const db = ctx.executor
 
-  {
-    const journalStore = createJournalStore(db)
-    const snapshot = await readSnapshot(metaDir)
-    const tableScope = resolveTableScope(tableSelector, tableKeysFromDefinitions(snapshot?.definitions ?? []))
+  const journalStore = createJournalStore(db)
+  const snapshot = await readSnapshot(metaDir)
+  const tableScope = resolveTableScope(tableSelector, tableKeysFromDefinitions(snapshot?.definitions ?? []))
 
-    await pluginRuntime.runOnConfigLoaded({
-      command: 'migrate',
-      config,
-      configPath,
-      tableScope,
-      flags,
-    })
+  await pluginRuntime.runOnConfigLoaded({
+    command: 'migrate',
+    config,
+    configPath,
+    tableScope,
+    flags,
+  })
 
-    await mkdir(migrationsDir, { recursive: true })
-    const files = await listMigrations(migrationsDir)
-    const journal = await journalStore.readJournal()
-    const appliedNames = new Set(journal.applied.map((entry) => entry.name))
-    const pendingAll = files.filter((f) => !appliedNames.has(f))
-    debug('migrate', `migrations: total=${files.length}, applied=${journal.applied.length}, pending=${pendingAll.length}`)
-    const checksumMismatches = await findChecksumMismatches(migrationsDir, journal)
-    if (checksumMismatches.length > 0) {
-      debug('migrate', `checksum mismatches: ${checksumMismatches.map((m) => m.name).join(', ')}`)
+  await mkdir(migrationsDir, { recursive: true })
+  const files = await listMigrations(migrationsDir)
+  const journal = await journalStore.readJournal()
+  const appliedNames = new Set(journal.applied.map((entry) => entry.name))
+  const pendingAll = files.filter((f) => !appliedNames.has(f))
+  debug('migrate', `migrations: total=${files.length}, applied=${journal.applied.length}, pending=${pendingAll.length}`)
+  const checksumMismatches = await findChecksumMismatches(migrationsDir, journal)
+  if (checksumMismatches.length > 0) {
+    debug('migrate', `checksum mismatches: ${checksumMismatches.map((m) => m.name).join(', ')}`)
+  }
+
+  if (checksumMismatches.length > 0) {
+    if (jsonMode) {
+      emitJson('migrate', {
+        mode: executeRequested ? 'execute' : 'plan',
+        scope: tableScope,
+        error: 'Checksum mismatch detected on applied migrations',
+        checksumMismatches,
+      })
+      process.exitCode = 1
+      return
     }
+    throw new Error(
+      `Checksum mismatch detected on applied migrations: ${checksumMismatches
+        .map((item) => item.name)
+        .join(', ')}`
+    )
+  }
 
-    if (checksumMismatches.length > 0) {
-      if (jsonMode) {
-        emitJson('migrate', {
-          mode: executeRequested ? 'execute' : 'plan',
-          scope: tableScope,
-          error: 'Checksum mismatch detected on applied migrations',
-          checksumMismatches,
-        })
-        process.exitCode = 1
-        return
-      }
-      throw new Error(
-        `Checksum mismatch detected on applied migrations: ${checksumMismatches
-          .map((item) => item.name)
-          .join(', ')}`
-      )
+  if (tableScope.enabled && tableScope.matchCount === 0) {
+    if (jsonMode) {
+      emitJson('migrate', {
+        mode: executeRequested ? 'execute' : 'plan',
+        scope: tableScope,
+        pending: [],
+        applied: [],
+        warning: `No tables matched selector "${tableScope.selector ?? ''}".`,
+      })
+    } else {
+      console.log(`No tables matched selector "${tableScope.selector ?? ''}". No migrations selected.`)
     }
+    return
+  }
 
-    if (tableScope.enabled && tableScope.matchCount === 0) {
-      if (jsonMode) {
-        emitJson('migrate', {
-          mode: executeRequested ? 'execute' : 'plan',
-          scope: tableScope,
-          pending: [],
-          applied: [],
-          warning: `No tables matched selector "${tableScope.selector ?? ''}".`,
-        })
-      } else {
-        console.log(`No tables matched selector "${tableScope.selector ?? ''}". No migrations selected.`)
+  const pending = tableScope.enabled
+    ? await filterPendingByScope(migrationsDir, pendingAll, new Set(tableScope.matchedTables))
+    : pendingAll
+
+  if (pending.length === 0) {
+    if (jsonMode) {
+      emitJson('migrate', {
+        mode: executeRequested ? 'execute' : 'plan',
+        scope: tableScope,
+        pending: [],
+        applied: [],
+      })
+    } else {
+      console.log('No pending migrations.')
+    }
+    return
+  }
+
+  const planned = {
+    mode: executeRequested ? 'execute' : 'plan',
+    scope: tableScope,
+    pending,
+  }
+
+  if (jsonMode && !executeRequested) {
+    emitJson('migrate', planned)
+    return
+  }
+
+  if (!jsonMode) {
+    if (tableScope.enabled) {
+      console.log(`Table scope: ${tableScope.selector ?? ''} (${tableScope.matchCount} matched)`)
+      for (const table of tableScope.matchedTables) console.log(`- ${table}`)
+    }
+    console.log(`Pending migrations: ${pending.length}`)
+    for (const file of pending) console.log(`- ${file}`)
+  }
+
+  let shouldExecute = executeRequested
+  if (!shouldExecute) {
+    if (isBackgroundOrCI() || jsonMode) {
+      if (!jsonMode) {
+        console.log('\nPlan only. Re-run with --apply to apply and journal these migrations.')
       }
       return
     }
 
-    const pending = tableScope.enabled
-      ? await filterPendingByScope(migrationsDir, pendingAll, new Set(tableScope.matchedTables))
-      : pendingAll
-
-    if (pending.length === 0) {
-      if (jsonMode) {
-        emitJson('migrate', {
-          mode: executeRequested ? 'execute' : 'plan',
-          scope: tableScope,
-          pending: [],
-          applied: [],
-        })
-      } else {
-        console.log('No pending migrations.')
-      }
-      return
+    const rl = createInterface({ input: process.stdin, output: process.stdout })
+    try {
+      console.log('')
+      console.log('Type "yes" to continue. Any other input cancels.')
+      const response = await rl.question('Apply pending migrations now? [no/yes]: ')
+      shouldExecute = response.trim().toLowerCase() === 'yes'
+    } finally {
+      rl.close()
     }
 
-    const planned = {
-      mode: executeRequested ? 'execute' : 'plan',
-      scope: tableScope,
-      pending,
-    }
-
-    if (jsonMode && !executeRequested) {
-      emitJson('migrate', planned)
-      return
-    }
-
-    if (!jsonMode) {
-      if (tableScope.enabled) {
-        console.log(`Table scope: ${tableScope.selector ?? ''} (${tableScope.matchCount} matched)`)
-        for (const table of tableScope.matchedTables) console.log(`- ${table}`)
-      }
-      console.log(`Pending migrations: ${pending.length}`)
-      for (const file of pending) console.log(`- ${file}`)
-    }
-
-    let shouldExecute = executeRequested
     if (!shouldExecute) {
-      if (isBackgroundOrCI() || jsonMode) {
-        if (!jsonMode) {
-          console.log('\nPlan only. Re-run with --apply to apply and journal these migrations.')
-        }
-        return
-      }
-
-      const rl = createInterface({ input: process.stdin, output: process.stdout })
-      try {
-        console.log('')
-        console.log('Type "yes" to continue. Any other input cancels.')
-        const response = await rl.question('Apply pending migrations now? [no/yes]: ')
-        shouldExecute = response.trim().toLowerCase() === 'yes'
-      } finally {
-        rl.close()
-      }
-
-      if (!shouldExecute) {
-        console.log('Migration apply cancelled by user.')
-        return
-      }
+      console.log('Migration apply cancelled by user.')
+      return
     }
+  }
 
-    const destructiveMigrations: string[] = []
-    const destructiveOperations: DestructiveOperationMarker[] = []
-    for (const file of pending) {
-      const sql = await readFile(join(migrationsDir, file), 'utf8')
-      if (migrationContainsDangerOperation(sql)) {
-        destructiveMigrations.push(file)
-        destructiveOperations.push(...collectDestructiveOperationMarkers(file, sql))
-      }
+  const destructiveMigrations: string[] = []
+  const destructiveOperations: DestructiveOperationMarker[] = []
+  for (const file of pending) {
+    const sql = await readFile(join(migrationsDir, file), 'utf8')
+    if (migrationContainsDangerOperation(sql)) {
+      destructiveMigrations.push(file)
+      destructiveOperations.push(...collectDestructiveOperationMarkers(file, sql))
     }
+  }
 
-    let destructiveAllowed = allowDestructive || config.safety?.allowDestructive === true
-    if (destructiveMigrations.length > 0 && !destructiveAllowed) {
-      const error =
-        'Blocked destructive migration execution. Re-run with --allow-destructive or set safety.allowDestructive=true after review.'
-      if (jsonMode) {
-        emitJson('migrate', {
-          mode: 'execute',
-          scope: tableScope,
-          error,
-          destructiveMigrations,
-          destructiveOperations,
-        })
-        process.exitCode = 3
-        return
-      }
-
-      if (isBackgroundOrCI()) {
-        printDestructiveOperationDetails(destructiveOperations)
-        throw new Error(
-          `${error}\nDestructive migrations: ${destructiveMigrations.join(', ')}\n` +
-            'Non-interactive run detected. Pass --allow-destructive to proceed.'
-        )
-      }
-
-      const confirmed = await confirmDestructiveExecution(destructiveOperations)
-      if (!confirmed) {
-        throw new Error(
-          `Destructive migration cancelled by user.\nDestructive migrations: ${destructiveMigrations.join(', ')}`
-        )
-      }
-      destructiveAllowed = true
-    }
-
-    if (destructiveMigrations.length > 0 && !destructiveAllowed) {
-      throw new Error('Blocked destructive migration execution.')
-    }
-
-    const appliedNow: MigrationJournalEntry[] = []
-
-    for (const file of pending) {
-      debug('migrate', `applying ${file}`)
-      const fullPath = join(migrationsDir, file)
-      const sql = await readFile(fullPath, 'utf8')
-      const parsedStatements = extractExecutableStatements(sql)
-      const operationSummaries = extractMigrationOperationSummaries(sql)
-      debug('migrate', `${file}: ${parsedStatements.length} statements, ${operationSummaries.length} operations`)
-      const statements = await pluginRuntime.runOnBeforeApply({
-        command: 'migrate',
-        config,
-        tableScope,
-        flags,
-        migration: file,
-        sql,
-        statements: parsedStatements,
-      })
-      for (let i = 0; i < statements.length; i++) {
-        const statement = statements[i] as string
-        await db.command(statement)
-        const operation = operationSummaries[i]
-        if (operation) {
-          await waitForDDLPropagation(db, operation.type, operation.key)
-        }
-      }
-
-      const entry: MigrationJournalEntry = {
-        name: file,
-        appliedAt: new Date().toISOString().replace('Z', ''),
-        checksum: checksumSQL(sql),
-      }
-      appliedNow.push(entry)
-      await journalStore.appendEntry(entry)
-      await pluginRuntime.runOnAfterApply({
-        command: 'migrate',
-        config,
-        tableScope,
-        flags,
-        migration: file,
-        statements,
-        appliedAt: entry.appliedAt,
-      })
-
-      if (!jsonMode) console.log(`Applied: ${file}`)
-    }
-
+  let destructiveAllowed = allowDestructive || config.safety?.allowDestructive === true
+  if (destructiveMigrations.length > 0 && !destructiveAllowed) {
+    const error =
+      'Blocked destructive migration execution. Re-run with --allow-destructive or set safety.allowDestructive=true after review.'
     if (jsonMode) {
       emitJson('migrate', {
         mode: 'execute',
         scope: tableScope,
-        applied: appliedNow,
+        error,
+        destructiveMigrations,
+        destructiveOperations,
       })
+      process.exitCode = 3
       return
     }
 
-    console.log(`\nMigrations recorded in ClickHouse _chkit_migrations table.`)
+    if (isBackgroundOrCI()) {
+      printDestructiveOperationDetails(destructiveOperations)
+      throw new Error(
+        `${error}\nDestructive migrations: ${destructiveMigrations.join(', ')}\n` +
+          'Non-interactive run detected. Pass --allow-destructive to proceed.'
+      )
+    }
+
+    const confirmed = await confirmDestructiveExecution(destructiveOperations)
+    if (!confirmed) {
+      throw new Error(
+        `Destructive migration cancelled by user.\nDestructive migrations: ${destructiveMigrations.join(', ')}`
+      )
+    }
+    destructiveAllowed = true
   }
+
+  if (destructiveMigrations.length > 0 && !destructiveAllowed) {
+    throw new Error('Blocked destructive migration execution.')
+  }
+
+  const appliedNow: MigrationJournalEntry[] = []
+
+  for (const file of pending) {
+    debug('migrate', `applying ${file}`)
+    const fullPath = join(migrationsDir, file)
+    const sql = await readFile(fullPath, 'utf8')
+    const parsedStatements = extractExecutableStatements(sql)
+    const operationSummaries = extractMigrationOperationSummaries(sql)
+    debug('migrate', `${file}: ${parsedStatements.length} statements, ${operationSummaries.length} operations`)
+    const statements = await pluginRuntime.runOnBeforeApply({
+      command: 'migrate',
+      config,
+      tableScope,
+      flags,
+      migration: file,
+      sql,
+      statements: parsedStatements,
+    })
+    for (let i = 0; i < statements.length; i++) {
+      const statement = statements[i] as string
+      await db.command(statement)
+      const operation = operationSummaries[i]
+      if (operation) {
+        await waitForDDLPropagation(db, operation.type, operation.key)
+      }
+    }
+
+    const entry: MigrationJournalEntry = {
+      name: file,
+      appliedAt: new Date().toISOString().replace('Z', ''),
+      checksum: checksumSQL(sql),
+    }
+    appliedNow.push(entry)
+    await journalStore.appendEntry(entry)
+    await pluginRuntime.runOnAfterApply({
+      command: 'migrate',
+      config,
+      tableScope,
+      flags,
+      migration: file,
+      statements,
+      appliedAt: entry.appliedAt,
+    })
+
+    if (!jsonMode) console.log(`Applied: ${file}`)
+  }
+
+  if (jsonMode) {
+    emitJson('migrate', {
+      mode: 'execute',
+      scope: tableScope,
+      applied: appliedNow,
+    })
+    return
+  }
+
+  console.log(`\nMigrations recorded in ClickHouse _chkit_migrations table.`)
 }
