--- packages/cli/src/bin/lib.ts
@@ -1,5 +1,0 @@
-export { CLI_VERSION } from './version.js'
-
-export { readSnapshot, summarizePlan } from './migration-store.js'
-export { emitJson } from './json-output.js'
-export { loadSchemaDefinitions } from './schema-loader.js'
