+++ packages/cli/src/test/check.e2e.test.ts
