+++ packages/cli/src/test/generate.e2e.test.ts
