+++ packages/cli/src/test/migrate.e2e.test.ts
