+++ packages/cli/src/test/migration-scenario.test.ts
