+++ packages/cli/src/test/plugin.test.ts
