+++ packages/cli/src/test/status.e2e.test.ts
