+++ packages/cli/src/test/table-scope.e2e.test.ts
