+++ packages/cli/tsconfig.json
@@ -5,5 +5,5 @@
     "outDir": "dist"
   },
   "include": ["src/**/*.ts"],
-  "exclude": ["src/**/*.test.ts", "src/e2e-testkit.ts"]
+  "exclude": ["src/test/**"]
 }
